feat: add extreme label for chart
